pulp-upload: Skip upload of existing RPMs on Pulp
Previously, all RPMs were uploaded regardless of whether
they already existed on Pulp. While Pulp doesn't update
the RPMs with the same checksum, this still resulted in
unnecessary upload overhead.

This deduplicates the uploads by checking the sha256 checksum
and upload only those RPMs that doesn't exist on Pulp, reducing
the upload time with redundant requests.

Note: This checks the checksum of the file to be uploaded with the
pkgID of the RPM on Pulp as pkgId and sha256sum are same on Pulp
and Pulp API offers checking a list of pkgIDs in a request.