pulp-upload: Existing RPMs are included in results json
RPMs that doesn't exist in a namespace on Pulp are only
uploaded to Pulp. They might have a different build ID
than the current one. Hence, they will be missing from
the list of artifacts for the current build, therefore in
the results json.

So, this includes all the existing RPMs in the results json,
that may be used by the relese pipline to promote contents.